update KF, EKF, UKF
Currently in roll_out.m, the measurement function is assumed to be dependent on the current control -- this might create a problem when the control is measurement-feedback. In future updates, we either 1) make the measurement only depend on the state; or 2) make the measurement depend on the past control instead of current control.